+++ src/components/shared/avatar/Avatar.js
@@ -25,7 +25,7 @@ const basetyles = {
         zIndex: 0
     },
     avatar: {
-        boxShadow: '0 0 10px 1px rgba(0,0,0,0.6)'
+        // boxShadow: '0 0 10px 1px rgba(0,0,0,0.6)'
     },
     innerImg: {
         width:'100%' ,height:'100%'
