+++ src/components/shared/avatar/index.js
@@ -1,0 +1,2 @@
+import Avatar from './Avatar';
+export default Avatar;
