+++ src/components/user/Grid/index.js
@@ -1,0 +1,2 @@
+import UsersGrid from './UsersGrid';
+export default UsersGrid;
