+++ src/components/user/Table/index.js
@@ -1,0 +1,2 @@
+import UsersTable from './UsersTable';
+export default UsersTable;
