+++ src/components/user/Table/styled/Container.js
@@ -1,0 +1,17 @@
+import styled from 'styled-components';
+
+const container = styled.div`
+    margin: 2%;
+    width: 80%
+`;
+
+const iconContainer = styled.div`
+    display: flex;
+    background-color: #90B7B6;
+    justify-content: center;
+    align-items: center;
+    height: 100%;
+`;
+
+export const Container = container;
+export const IconContainer = iconContainer;
