+++ src/components/user/Table/styled/styles.js
@@ -1,0 +1,24 @@
+const Styles = {
+    wrapper: {
+      width: '80%',
+      maxHeight:'440px',
+      margin: '2%',
+      overflow: 'auto'
+    },
+    avatarColumn: {
+      width: '6%',
+      minWidth: '0'
+    },
+    iconColumn: {
+      width: '3%',
+      // backgroundColor:'#90B7B6',
+      padding:0,
+      minWidth: '0'
+    },
+    iconColumnHeader: {
+      width: '3%',
+      padding:0    
+    }
+  };
+
+  export default Styles;
